Revert "chore(role): cache element list by role (#29130)"
This reverts commit 1ce3ca25a2676ca773a736035699cf3fe0c35d0d.

Added a regression test.
